+++ platform-dao/src/main/java/ua/com/fielden/platform/ioc/BasicWebServerModule.java
@@ -32,11 +32,9 @@ import ua.com.fielden.platform.entity.IEntityDeleteAction;
 import ua.com.fielden.platform.entity.IEntityEditAction;
 import ua.com.fielden.platform.entity.IEntityNavigationAction;
 import ua.com.fielden.platform.entity.IEntityNewAction;
-import ua.com.fielden.platform.entity.IReferenceHierarchy;
 import ua.com.fielden.platform.entity.ISecurityMatrixInsertionPoint;
 import ua.com.fielden.platform.entity.ISecurityMatrixSaveAction;
 import ua.com.fielden.platform.entity.ISecurityTokenTreeNodeEntity;
-import ua.com.fielden.platform.entity.ReferenceHierarchyDao;
 import ua.com.fielden.platform.entity.SecurityMatrixInsertionPointDao;
 import ua.com.fielden.platform.entity.SecurityMatrixSaveActionDao;
 import ua.com.fielden.platform.entity.SecurityTokenTreeNodeEntityDao;
@@ -69,6 +67,8 @@ import ua.com.fielden.platform.menu.ModuleDao;
 import ua.com.fielden.platform.menu.ModuleMenuItemDao;
 import ua.com.fielden.platform.menu.ViewDao;
 import ua.com.fielden.platform.menu.WebMenuItemInvisibilityDao;
+import ua.com.fielden.platform.ref_hierarchy.IReferenceHierarchy;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyDao;
 import ua.com.fielden.platform.security.IAuthorisationModel;
 import ua.com.fielden.platform.security.ISecurityRoleAssociationBatchAction;
 import ua.com.fielden.platform.security.IUserAndRoleAssociationBatchAction;
