+++ platform-dao/src/test/java/ua/com/fielden/platform/dao/CommonEntityDaoCompanionInstantiationTest.java
@@ -1,28 +1,39 @@
 package ua.com.fielden.platform.dao;
 
 import static java.lang.String.format;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import java.util.Random;
+import java.util.Set;
 
 import org.junit.Test;
 
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchy;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceLevelHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.ReferencedByLevelHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.TypeLevelHierarchyEntry;
 import ua.com.fielden.platform.test.PlatformTestDomainTypes;
 import ua.com.fielden.platform.test_config.AbstractDaoTestCase;
+import ua.com.fielden.platform.utils.CollectionUtil;
 
 public class CommonEntityDaoCompanionInstantiationTest extends AbstractDaoTestCase {
+    
+    private static final Set<Class<? extends AbstractEntity<?>>> types = CollectionUtil.setOf(ReferenceHierarchy.class, ReferenceHierarchyEntry.class, TypeLevelHierarchyEntry.class, ReferencedByLevelHierarchyEntry.class, ReferenceLevelHierarchyEntry.class);
 
+    
     @Test
     public void companion_objects_for_any_registered_domain_entity_can_be_instantiated_through_co_API_of_random_companion() {
         final Random rnd = new Random();
         final IEntityDao<?> randomCo = co$(PlatformTestDomainTypes.entityTypes.get(rnd.nextInt(PlatformTestDomainTypes.entityTypes.size())));
-        
-        for (Class<? extends AbstractEntity<?>> type: PlatformTestDomainTypes.entityTypes) {
+        PlatformTestDomainTypes.entityTypes.stream().filter(type -> !types.contains(type)).forEach(type -> {
             final IEntityDao<?> co = randomCo.co$(type);
             assertNotNull(format("Companion object for entity [%s] could not have been instantiated.", type.getName()), co);
-        }
+        });
     }
 
     @Test
@@ -30,11 +41,11 @@ public class CommonEntityDaoCompanionInstantiationTest extends AbstractDaoTestCa
         final Random rnd = new Random();
         final IEntityDao<?> randomCo = co$(PlatformTestDomainTypes.entityTypes.get(rnd.nextInt(PlatformTestDomainTypes.entityTypes.size())));
         
-        for (Class<? extends AbstractEntity<?>> type: PlatformTestDomainTypes.entityTypes) {
+        PlatformTestDomainTypes.entityTypes.stream().filter(type -> !types.contains(type)).forEach(type -> {
             final IEntityDao<?> co1 = randomCo.co$(type);
             final IEntityDao<?> co2 = randomCo.co$(type);
             assertTrue(format("Companion object for entity [%s] was not cached.", type.getName()), co1 == co2);
-        }
+        });
     }
 
     @Test
@@ -47,11 +58,11 @@ public class CommonEntityDaoCompanionInstantiationTest extends AbstractDaoTestCa
         
         assertFalse(randomCo1 == randomCo2);
         
-        for (Class<? extends AbstractEntity<?>> type: PlatformTestDomainTypes.entityTypes) {
+        PlatformTestDomainTypes.entityTypes.stream().filter(type -> !types.contains(type)).forEach(type -> {
             final IEntityDao<?> co1 = randomCo1.co$(type);
             final IEntityDao<?> co2 = randomCo2.co$(type);
             assertFalse(format("Companion objects for entity [%s] produced by different companions should not be the same.", type.getName()), co1 == co2);
-        }
+        });
     }
