+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/domain/PlatformDomainTypes.java
@@ -13,14 +13,9 @@ import ua.com.fielden.platform.entity.EntityEditAction;
 import ua.com.fielden.platform.entity.EntityExportAction;
 import ua.com.fielden.platform.entity.EntityNavigationAction;
 import ua.com.fielden.platform.entity.EntityNewAction;
-import ua.com.fielden.platform.entity.ReferenceHierarchy;
-import ua.com.fielden.platform.entity.ReferenceHierarchyEntry;
-import ua.com.fielden.platform.entity.ReferenceLevelHierarchyEntry;
-import ua.com.fielden.platform.entity.ReferencedByLevelHierarchyEntry;
 import ua.com.fielden.platform.entity.SecurityMatrixInsertionPoint;
 import ua.com.fielden.platform.entity.SecurityMatrixSaveAction;
 import ua.com.fielden.platform.entity.SecurityTokenTreeNodeEntity;
-import ua.com.fielden.platform.entity.TypeLevelHierarchyEntry;
 import ua.com.fielden.platform.entity.functional.centre.CentreContextHolder;
 import ua.com.fielden.platform.entity.functional.centre.SavingInfoHolder;
 import ua.com.fielden.platform.entity.functional.master.AcknowledgeWarnings;
@@ -39,6 +34,11 @@ import ua.com.fielden.platform.menu.WebMenuItemInvisibility;
 import ua.com.fielden.platform.migration.MigrationError;
 import ua.com.fielden.platform.migration.MigrationHistory;
 import ua.com.fielden.platform.migration.MigrationRun;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchy;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceLevelHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.ReferencedByLevelHierarchyEntry;
+import ua.com.fielden.platform.ref_hierarchy.TypeLevelHierarchyEntry;
 import ua.com.fielden.platform.security.session.UserSession;
 import ua.com.fielden.platform.security.user.SecurityRoleAssociation;
 import ua.com.fielden.platform.security.user.SecurityTokenInfo;
@@ -131,6 +131,5 @@ public class PlatformDomainTypes {
         types.add(TypeLevelHierarchyEntry.class);
         types.add(ReferencedByLevelHierarchyEntry.class);
         types.add(ReferenceLevelHierarchyEntry.class);
-        types.add(MasterInfo.class);
     }
 }
