+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/ref_hierarchy/AbstractTreeEntry.java
@@ -1,9 +1,10 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.ref_hierarchy;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.annotation.IsProperty;
 import ua.com.fielden.platform.entity.annotation.Observable;
 import ua.com.fielden.platform.entity.annotation.Title;
