+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/ref_hierarchy/ReferencedByLevelHierarchyEntry.java
@@ -1,5 +1,6 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.ref_hierarchy;
 
+import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.annotation.DescTitle;
 import ua.com.fielden.platform.entity.annotation.IsProperty;
 import ua.com.fielden.platform.entity.annotation.KeyTitle;
