+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/ref_hierarchy/TypeLevelHierarchyEntry.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.ref_hierarchy;
 
 import ua.com.fielden.platform.entity.annotation.DescTitle;
 import ua.com.fielden.platform.entity.annotation.IsProperty;
