+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/hierarchy/ReferenceHierarchyMaster.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 import ua.com.fielden.platform.basic.IValueMatcherWithContext;
 import ua.com.fielden.platform.dom.DomElement;
 import ua.com.fielden.platform.dom.InnerTextElement;
-import ua.com.fielden.platform.entity.ReferenceHierarchy;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchy;
 import ua.com.fielden.platform.utils.ResourceLoader;
 import ua.com.fielden.platform.web.centre.api.actions.EntityActionConfig;
 import ua.com.fielden.platform.web.centre.api.actions.impl.EntityActionBuilder;
